Added 8 unit tests for FileSystemDatabaseManager: 6 passing, 2 failling due to bugs

+++ AcademicManager/DataBase.UnitTests/DatabaseManagerUnitTests.cs
@@ -1,5 +1,6 @@
 ﻿namespace DataBase.UnitTests
 {
+    using System;
     using System.IO;
     using Database.Implementations;
     using Database.Interfaces;
@@ -22,7 +23,10 @@
         [TearDown]
         public void TearDown()
         {
-            Directory.Delete(DatabaseFilePath, true);
+            if (Directory.Exists(DatabaseFilePath))
+            {
+                Directory.Delete(DatabaseFilePath, true);
+            }
         }
 
         private const string DatabaseFilePath = @"C:\Databases";
@@ -31,16 +35,77 @@
         private IDatabaseManager<FileSystemDatabase> databaseManager;
 
         [Test]
-        public void ShouldSuccesfullyOpenACreatedDatabase()
+        public void ShouldNotThrowWhenCallingCreateWithEmptyParams()
+        {
+            var actualDatabase = databaseManager.Create(string.Empty, string.Empty);
+            Assert.AreEqual(@"C:\Databases", actualDatabase.ConnectionString);
+            Assert.AreEqual(string.Empty, actualDatabase.Name);
+        }
+
+        [Test]
+        public void ShouldOpenedDatabaseNameAndConnectionStringShoudNotEndWithBackSlash()
+        {
+            databaseManager.Create(DatabaseFilePath, DatabaseName);
+            var actualDatabase = databaseManager.Open(databaseConnectionString);
+
+            Assert.IsNotNull(actualDatabase);
+            Assert.IsFalse(actualDatabase.ConnectionString.EndsWith("\\", StringComparison.Ordinal));
+            Assert.IsFalse(actualDatabase.Name.EndsWith("\\", StringComparison.Ordinal));
+        }
+
+        [Test]
+        public void ShouldSuccesfullyCreateANewDatabase()
         {
             var expectedDatabase = databaseManager.Create(DatabaseFilePath, DatabaseName);
 
             Assert.IsNotNull(expectedDatabase);
+            Assert.AreEqual(DatabaseFilePath, expectedDatabase.ConnectionString);
+            Assert.AreEqual(DatabaseName, expectedDatabase.Name);
+        }
 
+        [Test]
+        public void ShouldSuccesfullyOpenAnExistingDatabase()
+        {
+            databaseManager.Create(DatabaseFilePath, DatabaseName);
             var actualDatabase = databaseManager.Open(databaseConnectionString);
 
-            Assert.AreEqual(expectedDatabase.ConnectionString, actualDatabase.ConnectionString);
-            Assert.AreEqual(expectedDatabase.Name, actualDatabase.Name);
+            Assert.IsNotNull(actualDatabase);
+            Assert.AreEqual(DatabaseFilePath, actualDatabase.ConnectionString);
+            Assert.AreEqual(DatabaseName, actualDatabase.Name);
+        }
+
+        [Test]
+        [ExpectedException(typeof(FileNotFoundException))]
+        public void ShouldThrowWhenCallingOpenWithEmpty()
+        {
+            //BUG: does not throw expected error
+            databaseManager.Open(string.Empty);
+        }
+
+        [Test]
+        [ExpectedException(typeof(IOException))]
+        public void ShouldThrowWhenCreatingADatabaseWithTheSameConnectionString()
+        {
+            databaseManager.Create(DatabaseFilePath, DatabaseName);
+            databaseManager.Create(DatabaseFilePath, DatabaseName);
+        }
+
+        [Test]
+        [ExpectedException(typeof(IOException))]
+        public void ShouldThrowWhenCreatingADatabaseWithMalformedFilePath()
+        {
+            databaseManager.Create("not a filepath", DatabaseName);
+        }
+
+        [Test]
+        [ExpectedException(typeof(FileNotFoundException))]
+        public void ShouldThrowWhenOpeningANonExistingDatabase()
+        {
+            //BUG: this does not fail with the expected error because the implementation does not properly verify
+            //BUG: that the actual database folder exists, it checks that its parent folder exists instead
+            databaseManager.Create(DatabaseFilePath, DatabaseName);
+            var nonExistingDatabase = Path.Combine(DatabaseFilePath, DatabaseName) + "NotExists";
+            databaseManager.Open(nonExistingDatabase);
         }
     }
 }
